+++ deepfellow/server/utils/options.py
@@ -16,11 +16,12 @@ import typer
 
 from deepfellow.common.defaults import DF_SERVER_DIRECTORY
 from deepfellow.common.env import env_get, env_set
+from deepfellow.common.state import state
 
 
-def get_default_server_directory(ctx: typer.Context) -> Path:
+def get_default_server_directory() -> Path:
     """Return default server directory."""
-    config_file = ctx.obj.get("cli-config-file")
+    config_file = state.cli_config_file
 
     default_dir = env_get(config_file, "DF_DEFAULT_SERVER_DIR", should_raise=False)
 
@@ -30,9 +31,9 @@ def get_default_server_directory(ctx: typer.Context) -> Path:
     return Path(default_dir)
 
 
-def set_default_server_directory(ctx: typer.Context, directory: str | Path, force: bool = False) -> None:
+def set_default_server_directory(directory: str | Path, force: bool = False) -> None:
     """Sets default server directory."""
-    config_file = ctx.obj.get("cli-config-file")
+    config_file = state.cli_config_file
 
     if force:
         env_set(config_file, "DF_DEFAULT_SERVER_DIR", str(directory))
@@ -42,11 +43,11 @@ def set_default_server_directory(ctx: typer.Context, directory: str | Path, forc
             env_set(config_file, "DF_DEFAULT_SERVER_DIR", str(directory))
 
 
-def default_directory_callback(ctx: typer.Context, dir: str | None) -> Path:
+def default_directory_callback(dir: str | None) -> Path:
     """Return default server directory if str is None."""
     out_dir = DF_SERVER_DIRECTORY
     if dir is None:
-        config_file = ctx.obj.get("cli-config-file")
+        config_file = state.cli_config_file
         default_dir = env_get(config_file, "DF_DEFAULT_SERVER_DIR", should_raise=False)
         if default_dir:
             out_dir = Path(default_dir)